+++ package.json
@@ -22,6 +22,7 @@
     "@types/reach__router": "^1.3.4",
     "@types/react": "^16.9.32",
     "@types/react-dom": "^16.9.6",
+    "axios": "^0.19.2",
     "d3": "^5.16.0",
     "formik": "^2.1.4",
     "moment": "^2.24.0",
@@ -55,6 +56,7 @@
     ]
   },
   "devDependencies": {
+    "@types/yup": "^0.28.0",
     "@typescript-eslint/eslint-plugin": "^2.28.0",
     "@typescript-eslint/parser": "^2.28.0",
     "eslint": "^6.8.0",
